Check winning condition in every direction
